Revert of few other files for statcode style
