--- statcode/constants.py
@@ -1,18 +1,0 @@
-import os.path
-
-CURR_DIR = os.path.dirname(os.path.realpath(__file__))
-
-# Scroll actions
-SCROLL_LINE_UP = "line up"
-SCROLL_LINE_DOWN = "line down"
-SCROLL_PAGE_UP = "page up"
-SCROLL_PAGE_DOWN = "page down"
-SCROLL_TO_TOP = "to top"
-SCROLL_TO_END = "to end"
-
-# ASCII color codes
-YELLOW = '\033[33m'
-RED = "\033[31m"
-BOLD = '\033[1m'
-UNDERLINE = '\033[4m'
-END = "\033[0m"
